Convertiong Assigned to Me and Mentioned Me widgets to one common base class for them to work with JSON ConfigurationData strings.
